requirements.txt
Django==5.2.8
djangorestframework==3.16.1
django-cors-headers==4.9.0
dj-database-url==1.1.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
